requirements.txt
PyYAML
requests
python-dotenv

# DNS
git+https://github.com/paulc/dnslib.git

# Webhook server
Flask
waitress

# Git
GitPython